vsf/requirements.txt
amqp==2.6.1
asgiref==3.2.8
attrs==19.3.0
autopep8==1.5.3
billiard==3.6.3.0
celery==4.4.7
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-datatables-view==1.19.1
django-environ==0.4.5
django-filter==2.3.0
django-model-utils==4.0.0
django-sass==1.0.0
django-timezone-field==4.0
django-webpack-loader==0.7.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
drf-yasg==1.17.1
flower==0.9.5
gunicorn==20.0.4
humanize==2.6.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==4.6.11
libsass==0.20.1
Markdown==3.2.2
MarkupSafe==1.1.1
packaging==20.4
prometheus-client==0.8.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
swagger-spec-validator==2.7.3
toml==0.10.1
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
